servive: Rename service ID management functions
Also moves the kvstore specific code into id_kvstore.go so an alternative can
be provided.
